Remove PlayerContext and usePlayer hook
Redundant re-renders and bugs, was not providing much benefit. if the
playerId is tampered, the backend anyways has its own track of it. so
the core functionality won't be affected without it. However, needed for
UI elements where the current player is tracked. Think of another
workaround, something like involving the backend